feat(operators): add GateUpLinear op type, params, schema and serde
Introduce OpType::kGateUpLinear and GateUpLinearParams (gate/up out_features,
has_bias), wire them through the operator schema (19th entry, RuntimeOnly,
dual gate/up outputs), serde canonical form, ToString, and graph dump.
OpType narrows to uint8_t so the enum stays byte-packed; the backend_fwd.h
forward declaration and dispatch_design.md are updated to match. Bias inputs
are intentionally not representable by the two-input schema yet.

+++ include/aethermind/backend/backend_fwd.h
@@ -10,7 +10,7 @@ class BackendFactory;
 class BackendRegistry;
 class KernelRegistry;
 
-enum class OpType : uint16_t;
+enum class OpType : uint8_t;
 enum class IsaLevel : uint8_t;
 enum class ExecPhase : uint8_t;
 enum class WeightFormat : uint8_t;
